prepping for dj17 upgrade
oy the migrations has me all verklempt.  No, seriously - it might be
worth it to reorganize the app now, because the ./manage.py
makemigrations command does not seem to get the rooibos/appname
structure.  A lot of failing, a lot of hand holding, bleh.

Adding a temp "drop and remake database" sql script that's a dump of a
working django 1.6 database, because the new migrations are... vexing,
and I'm not sure that getting halfway through the migration doesn't do
something to the db.