messing around with stats views - req. django-tables2
`pip install django-tables2`
Except I'm not clear it's worth it - there's a long way to go to
replicate what took 15 min to implement in admin.statistics.activities,
and the only real payoff would be adding some charts or something. But
it basically works now. I'm going to commit the admin.py to my j16
branch since it should be in there anyway.